split organelle project into submodules
